+++ src/main/scala/edu/ie3/simona/model/participant/ParticipantModelShell.scala
@@ -182,7 +182,7 @@ final case class ParticipantModelShell[
       val (modelOp, modelNextTick) =
         model.determineOperatingPoint(currentState)
       // Sanity check
-      if (modelNextTick.exists(_ <= tick))
+      if modelNextTick.exists(_ <= tick) then
         throw new CriticalFailureException(
           s"$identifier: Next tick ($modelNextTick) is same as or earlier than the current tick ($tick)."
         )
@@ -279,7 +279,7 @@ final case class ParticipantModelShell[
 
     val updatedFlexOptions = flexType match {
       case FlexType.PowerLimit =>
-        if (operationInterval.includes(tick)) {
+        if operationInterval.includes(tick) then {
           flexModel.determineFlexOptions(currentState)
         } else {
           // Out of operation, there's no way to operate besides 0 kW
@@ -335,7 +335,7 @@ final case class ParticipantModelShell[
       determineOperatingPoint(modelOperatingPoint, currentTick)
 
     // Sanity check
-    if (newChangeIndicator.changesAtTick.exists(_ <= currentTick))
+    if newChangeIndicator.changesAtTick.exists(_ <= currentTick) then
       throw new CriticalFailureException(
         s"$identifier: Next tick (${newChangeIndicator.changesAtTick}) is same as or earlier than the current tick ($currentTick)."
       )
@@ -363,7 +363,7 @@ final case class ParticipantModelShell[
       modelOperatingPoint: () => (OP, OperationChangeIndicator),
       currentTick: Long,
   ): (OP, OperationChangeIndicator) = {
-    if (operationInterval.includes(currentTick)) {
+    if operationInterval.includes(currentTick) then {
       modelOperatingPoint()
     } else {
       // Current tick is outside of operation interval.
@@ -386,7 +386,7 @@ final case class ParticipantModelShell[
       currentTick: Long,
       nextDataTick: Option[Long],
   ): OperationChangeIndicator = {
-    if (operationInterval.includes(currentTick)) {
+    if operationInterval.includes(currentTick) then {
       // The next activation tick should be the earliest of
       // the next tick request by the model, the next data tick and
       // the end of the operation interval
@@ -441,7 +441,7 @@ final case class ParticipantModelShell[
   private def determineCurrentState(tick: Long): S = {
     // new state is only calculated if there's an old state and an operating point
     val newState =
-      if (state.tick < tick) {
+      if state.tick < tick then {
         // If the state is old, an operating point needs
         // to be present to determine the curren state
         model.determineState(
@@ -455,7 +455,7 @@ final case class ParticipantModelShell[
         state
       }
 
-    if (newState.tick != tick)
+    if newState.tick != tick then
       throw new CriticalFailureException(
         s"$identifier: The current state $newState is not set to current tick $tick"
       )